added editing rooms feature with leaving the site warning message

+++ client/src/components/addRoom/addImages/ImagesList.js
@@ -11,12 +11,14 @@ import deleteFile from '../../../firebase/deleteFile';
 
 const ImagesList = () => {
   const {
-    state: { images, currentUser },
+    state: { images, currentUser, updatedRoom },
     dispatch,
   } = useValue();
 
   const handleDelete = async (image) => {
     dispatch({ type: 'DELETE_IMAGE', payload: image });
+    if (updatedRoom)
+      return dispatch({ type: 'UPDATE_DELETED_IMAGES', payload: [image] });
     const imageName = image?.split(`${currentUser?.id}%2F`)[1]?.split('?')[0];
     try {
       await deleteFile(`rooms/${currentUser?.id}/${imageName}`);
